move fetcher to api store module
