+++ src/lib/store/index.ts
@@ -1,13 +1,6 @@
-import { writable } from 'svelte/store';
+import * as api from './api';
+import * as browser from './browser';
 
-const cache = new Map();
-
-export function fetcher(url: RequestInfo) {
-	const store = writable(new Promise(() => {}));
-	if (cache.has(url)) store.set(Promise.resolve(cache.get(url)));
-	const load = async () => {
-		const data = await fetch(url).then((r) => r.json());
-		cache.set(url, data), store.set(Promise.resolve(data));
-	};
-	return load(), store;
-}
+const store = { api, browser };
+export { api, browser };
+export default store;
